services/telegram-bot/requirements.txt
# Personal System Telegram Bot - Production Requirements
# Optimized for Coolify deployment

# Core Telegram Bot
python-telegram-bot==20.7

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0
cryptography==41.0.7
schedule==1.2.0
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# HTTP client (compatible with python-telegram-bot)
httpx>=0.25.0
aiohttp>=3.8.0

# Data processing (minimal for production)
pandas==2.1.4
numpy==1.24.3

# Voice processing and AI (essential only)
openai==1.3.0
pydub==0.25.1
SpeechRecognition==3.10.0

# Health check and monitoring
psutil==5.9.6

# Production optimizations
gunicorn==21.2.0
uvicorn==0.24.0

# Security
bcrypt==4.1.2

# Logging
structlog==23.2.0

# Optional: For advanced AI features (uncomment if needed)
# transformers==4.35.0
# torch==2.1.0
# sentence-transformers==2.2.2